Addeed Layout and Notes
