+++ src/components/Header/Header.css
@@ -1,71 +1,65 @@
-/* *{
-background-color: black;
-}
 .header {
     display: flex;
     justify-content: space-between;
     align-items: center;
-  
     padding: 10px 20px;
     border-bottom: gray solid 2px;
+}
 
-  }
-  
-  .header-logo h1 {
+.header-logo h1 {
     color: white;
     font-size: 24px;
-  }
-  
-  .header-nav ul {
+}
+
+.header-nav ul {
     list-style-type: none;
     display: flex;
     gap: 90px;
-  }
-  
-  .header-nav ul li {
+}
+
+.header-nav ul li {
     font-size: 18px;
-  }
-  
-  .header-nav ul li a {
+}
+
+.header-nav ul li a {
     color: white;
     text-decoration: none;
-  }
-  
-  .header-nav ul li a:hover {
+}
+
+.header-nav ul li a:hover {
     text-decoration: underline;
-  }
-  
-  .signin-btn {
+}
+
+.signin-btn {
     background-color: #fff;
-    color:black;
+    color: black;
     border: none;
     padding: 10px 20px;
     font-size: 16px;
     cursor: pointer;
-  }
-  
-  .signin-btn:hover {
-    background-color: #4CAF50;
+}
+
+.signin-btn:hover {
+    background-color: #4caf50;
     color: white;
-  }
-  
-  html {
+}
+
+html {
     scroll-behavior: smooth;
-  }
-  
-  .section {
+}
+
+.section {
     padding: 50px;
     margin: 20px;
     background-color: #f0f0f0;
     border-radius: 8px;
-  }
-  
-  .section h2 {
+}
+
+.section h2 {
     font-size: 24px;
-  }
-  
-  .section p {
+}
+
+.section p {
     font-size: 16px;
     color: #555;
-  }
-   */
+}
