+++ src/components/Header/Header.jsx
@@ -1,12 +1,10 @@
-// src/components/Header/Header.jsx
-import React from 'react';
-import './Header.css'; // CSS for styling
+import './Header.css';
 
 const Header = () => {
   return (
     <div>
       {/* Header with Navigation Links */}
-      <header className="header">
+      <header className="header bg-[#27272a]">
         <div className="header-logo">
           <h1>Edubuddy</h1>
         </div>
@@ -23,7 +21,7 @@ const Header = () => {
       </header>
 
       {/* Sections that will be scrolled to */}
-      
+
     </div>
   );
 };
