+++ src/constant/index.js
@@ -1,0 +1,1 @@
+export const Base_URL = "http://localhost:4000";
