+++ src/pages/notes/index.jsx
@@ -1,12 +1,12 @@
-import React from 'react'
 import Header from '../../components/Header/header'
 import PageLayout from '../../layouts'
 
 export default function Notes() {
+
   return (
     <>
-    <Header />
-      <PageLayout />
+      <Header />
+      <PageLayout title="Notes" />
     </>
   )
 }
